+++ src/components/signup/StepRenderer.tsx
@@ -5,6 +5,7 @@ import { useNavigate } from "react-router-dom";
 import { useRecoilState, useRecoilValue } from "recoil";
 import { stepNum } from "../../atom/signup/signup";
 import PasswordAgreeChecking from "./PasswordAgreeChecking";
+import WelcomeSignup from "../welcomeSignup/WelcomeSignup";
 
 export default function StepRenderer() {
   const step = useRecoilValue(stepNum);
@@ -14,7 +15,8 @@ export default function StepRenderer() {
       navigate(-1);
       break;
     case 1:
-      return <Role />;
+      return <WelcomeSignup />;
+    // return <Role />;
     case 2:
       return <NameEmail />;
     case 3:
