+++ src/components/welcomeSignup/AfterSignup.tsx
@@ -1,17 +1,24 @@
 import { styled } from "styled-components";
 import { TuticeWithTextCommonIc } from "../../assets";
 import SignupTitleLayout from "../signup/SignupTitleLayout";
+import ButtonLayout from "./ButtonLayout";
 
 export default function AfterSignup() {
   const MAIN_TEXT = `이은수 선생님 환영합니다! \n 수업 나무 관리를 통해 \n 과외 관리를 시작해보세요. `;
 
   const SUB_TEXT = "*수업나무 : 수강생을 추가하여 관리할 수 있는 서비스입니다.";
+
+  function handleToSetAlert() {}
   return (
-    <Container>
-      <TuticeWithTextCommonIcon />
-      <SignupTitleLayout MainText={MAIN_TEXT} />
-      <SubText>{SUB_TEXT}</SubText>
-    </Container>
+    <>
+      <Container>
+        <TuticeWithTextCommonIcon />
+        <SignupTitleLayout MainText={MAIN_TEXT} />
+        <SubText>{SUB_TEXT}</SubText>
+      </Container>
+
+      <ButtonLayout onClick={handleToSetAlert} buttonText={"수업 나무 생성"} />
+    </>
   );
 }
